Code review cleanups
